Tests/LibWeb/Text/expected/Worker/Worker-echo.txt

Got message from worker: "loaded"
Got message from worker: {"msg":"marco"}
DONE
